backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
google-auth>=2.30
openai>=1.40
python-dotenv>=1.0
PyYAML>=6.0
